refactor: split chilldkg.rs into submodules
replaces 724e30fb63478dff3a9ed0db1dc4d8acd21f0d19
