+++ schnorr_fun/src/frost/chilldkg.rs
@@ -33,19 +33,13 @@
 //! - [`certpedpop`]: `encpedpop` where each party also certifies the output so they can cryptographically convince each other that the key generation was successful.
 //!
 //! [ChillDKG]: https://github.com/BlockstreamResearch/bip-frost-dkg
-use crate::{Schnorr, frost::*};
-use alloc::{
-    collections::{BTreeMap, BTreeSet},
-    vec::Vec,
-};
-use secp256kfun::{
-    KeyPair,
-    hash::{Hash32, HashAdd},
-    nonce::NonceGen,
-    poly,
-    prelude::*,
-    rand_core,
-};
+
+pub mod certpedpop;
+pub mod encpedpop;
+pub mod simplepedpop;
+
+use crate::Schnorr;
+use secp256kfun::{KeyPair, hash::Hash32, nonce::NonceGen, prelude::*};
 
 /// A trait for signature schemes that can be used to certify the DKG output.
 ///
@@ -147,1338 +141,10 @@ pub mod vrf_cert {
     }
 }
 
-/// SimplePedPop is a bare bones secure distributed key generation algorithm that leaves a lot left
-/// up to the application.
-///
-/// The application must figure out:
-///
-/// - How to secretly transport secret share contribution from each contributor to their intended destination
-/// - Checking that each party got the correct output by comparing [`AggKeygenInput::cert_bytes`] on each of them.
-///
-/// [`AggKeygenInput::cert_bytes`]: simplepedpop::AggKeygenInput::cert_bytes
-pub mod simplepedpop {
-    use super::*;
-    use crate::{Message, Signature};
-    use alloc::{
-        collections::{BTreeMap, BTreeSet},
-        vec::Vec,
-    };
-    use secp256kfun::hash::Hash32;
-
-    /// A party that generates secret input to the key generation. You need at least one of these
-    /// and if at least one of these parties is honest then the final secret key will not be known by an
-    /// attacker (unless they obtain `t` shares!).
-    #[derive(Clone, Debug, PartialEq)]
-    #[cfg_attr(feature = "bincode", derive(bincode::Encode, bincode::Decode))]
-    #[cfg_attr(
-        feature = "serde",
-        derive(crate::fun::serde::Deserialize, crate::fun::serde::Serialize),
-        serde(crate = "crate::fun::serde")
-    )]
-    pub struct Contributor {
-        my_key_contrib: Point,
-        my_index: u32,
-    }
-
-    impl Contributor {
-        /// Generates the keygen input for a party at `my_index`. Note that `my_index`
-        /// has nothing to do with the "receiver" index (the `ShareIndex` of share receivers). If
-        /// there are `n` `KeyGenInputParty`s then each party must be assigned an index from `0` to `n-1`.
-        ///
-        /// This method return `Self` to retain the state of the protocol which is needded to verify
-        /// the aggregated input later on.
-        pub fn gen_keygen_input<H, NG>(
-            schnorr: &Schnorr<H, NG>,
-            threshold: u32,
-            share_receivers: &BTreeSet<ShareIndex>,
-            my_index: u32,
-            rng: &mut impl rand_core::RngCore,
-        ) -> (Self, KeygenInput, SecretKeygenInput)
-        where
-            H: Hash32,
-            NG: NonceGen,
-        {
-            let secret_poly = poly::scalar::generate(threshold as usize, rng);
-            let pop_keypair = KeyPair::new_xonly(secret_poly[0]);
-            // XXX The thing that's singed differs from the spec
-            let pop = schnorr.sign(&pop_keypair, Message::empty());
-            let com = poly::scalar::to_point_poly(&secret_poly);
-
-            let shares = share_receivers
-                .iter()
-                .map(|index| (*index, poly::scalar::eval(&secret_poly, *index)))
-                .collect();
-            let self_ = Self {
-                my_key_contrib: com[0],
-                my_index,
-            };
-            let msg = KeygenInput { com, pop };
-            (self_, msg, shares)
-        }
-
-        /// Verifies that the coordinator has honestly included this party's input into the
-        /// aggregated input.
-        ///
-        /// This passing by itself doesn't mean that the key generation was successful. All
-        /// `Contributor`s must agree on this fact and all parties must have received the same
-        /// `AggKeygenInput` and validated it.
-        pub fn verify_agg_input(
-            self,
-            agg_input: &AggKeygenInput,
-        ) -> Result<(), ContributionDidntMatch> {
-            let my_got_contrib = agg_input
-                .key_contrib
-                .get(self.my_index as usize)
-                .map(|(point, _)| *point);
-            let my_expected_contrib = self.my_key_contrib;
-            if Some(my_expected_contrib) != my_got_contrib {
-                return Err(ContributionDidntMatch);
-            }
-
-            Ok(())
-        }
-    }
-
-    /// Produced by [`Contributor::gen_keygen_input`]. This is sent from the each
-    /// `Contributor` to the *coordinator*.
-    #[cfg_attr(feature = "bincode", derive(bincode::Encode, bincode::Decode))]
-    #[cfg_attr(
-        feature = "serde",
-        derive(crate::fun::serde::Deserialize, crate::fun::serde::Serialize),
-        serde(crate = "crate::fun::serde")
-    )]
-    #[derive(Clone, Debug, PartialEq)]
-    pub struct KeygenInput {
-        /// The polynomial commitment of the contributor.
-        pub com: Vec<Point>,
-        /// Their proof-of-possession signature on the first coefficient.
-        pub pop: Signature,
-    }
-
-    /// Map from share index to secret share contribution from the [`Contributor`].
-    ///
-    /// Each entry in the map must be sent to the corresponding party.
-    pub type SecretKeygenInput = BTreeMap<ShareIndex, Scalar<Secret, Zero>>;
-
-    /// Stores the state of the coordinator as it aggregates inputs from [`Contributor`]s.
-    #[derive(Clone, Debug, PartialEq)]
-    pub struct Coordinator {
-        threshold: u32,
-        inputs: BTreeMap<u32, Option<KeygenInput>>,
-    }
-
-    impl Coordinator {
-        /// Creates a new coordinator with:
-        ///
-        /// - `threshold`: of key we're trying to generate
-        /// - `n_contributors`: The number of [`Contributor`]s
-        pub fn new(threshold: u32, n_contributors: u32) -> Self {
-            assert!(threshold > 0);
-            Self {
-                threshold,
-                inputs: (0..n_contributors).map(|i| (i, None)).collect(),
-            }
-        }
-
-        /// Adds an `input` from a [`Contributor`].
-        ///
-        /// Note verifying this is the correct input from the correct party is up to your application!
-        pub fn add_input<H: Hash32, NG>(
-            &mut self,
-            schnorr: &Schnorr<H, NG>,
-            from: u32,
-            input: KeygenInput,
-        ) -> Result<(), &'static str> {
-            let entry = match self.inputs.get_mut(&from) {
-                Some(maybe_input) => match maybe_input {
-                    Some(_) => return Err("we already have input from this party"),
-                    none => none,
-                },
-                None => return Err("no input expected from this party"),
-            };
-            if input.com.len() != self.threshold as usize {
-                return Err("input has the wrong threshold");
-            }
-
-            let (first_coeff_even_y, _) = input.com[0].into_point_with_even_y();
-            if !schnorr.verify(&first_coeff_even_y, Message::empty(), &input.pop) {
-                return Err("☠ pop didn't verify");
-            }
-            *entry = Some(input);
-
-            Ok(())
-        }
-
-        /// Which [`Contributor`]s are we missing input from.
-        pub fn missing_from(&self) -> BTreeSet<u32> {
-            self.inputs
-                .iter()
-                .filter_map(|(index, input)| match input {
-                    None => Some(*index),
-                    Some(_) => None,
-                })
-                .collect()
-        }
-
-        /// Has the coordinator received input from each [`Contributor`].
-        pub fn is_finished(&self) -> bool {
-            self.inputs.values().all(|v| v.is_some())
-        }
-
-        /// Try and finish input aggregation step.
-        ///
-        /// Returns `None` if [`is_finished`] returns `false`.
-        ///
-        /// [`is_finished`]: Self::is_finished
-        pub fn finish(self) -> Option<AggKeygenInput> {
-            if !self.is_finished() {
-                return None;
-            }
-            let inputs = self.inputs.into_values().flatten().collect::<Vec<_>>();
-            // The "key contributions" are separated out and treated specially since they can't be
-            // aggregated by the coordinator since each one needs to be validated against a
-            // proof-of-possesson.
-            let key_contrib = inputs
-                .iter()
-                .map(|message| (message.com[0], message.pop))
-                .collect();
-
-            // The rest of the coefficients can be aggregated
-            let mut agg_poly =
-                vec![Point::<NonNormal, Public, _>::zero(); self.threshold as usize - 1];
-            for message in inputs {
-                for (i, com) in message.com[1..].iter().enumerate() {
-                    agg_poly[i] += com
-                }
-            }
-
-            let agg_poly = poly::point::normalize(agg_poly).collect::<Vec<_>>();
-
-            Some(AggKeygenInput {
-                key_contrib,
-                agg_poly,
-            })
-        }
-    }
-
-    /// Key generation inputs after being aggregated by the coordinator
-    #[derive(Clone, Debug, PartialEq)]
-    #[cfg_attr(feature = "bincode", derive(bincode::Encode, bincode::Decode))]
-    #[cfg_attr(
-        feature = "serde",
-        derive(crate::fun::serde::Deserialize, crate::fun::serde::Serialize),
-        serde(crate = "crate::fun::serde")
-    )]
-    pub struct AggKeygenInput {
-        /// The key contribution from each [`Contributor`]
-        pub key_contrib: Vec<(Point, Signature)>,
-        /// The aggregated non-constant term polynomial
-        pub agg_poly: Vec<Point<Normal, Public, Zero>>,
-    }
-
-    impl AggKeygenInput {
-        /// Gets the `SharedKey` that this aggregated input produces.
-        ///
-        /// ## Security
-        ///
-        /// ⚠ Just because you can call this doesn't mean you can use the `SharedKey` securely yet!
-        ///
-        /// You have to have checked that all parties (contributors and receivers) think it's valid
-        /// *and* have the same copy first.
-        pub fn shared_key(&self) -> SharedKey<Normal, Zero> {
-            let public_key = self
-                .key_contrib
-                .iter()
-                .fold(Point::zero(), |agg, (point, _)| g!(agg + point))
-                .normalize();
-            let mut poly = self.agg_poly.clone();
-            poly.insert(0, public_key);
-            SharedKey::from_poly(poly)
-        }
-
-        /// The *certification* bytes. Checking all parties have the same output of this function is
-        /// enough to check they have the same `AggKeygenInput`.
-        ///
-        /// In `simplepedpop` this is just the coefficients of the polynomial.
-        pub fn cert_bytes(&self) -> Vec<u8> {
-            let mut cert_bytes = vec![];
-            cert_bytes.extend((self.agg_poly.len() as u32).to_be_bytes());
-            for coeff in self.shared_key().point_polynomial() {
-                cert_bytes.extend(coeff.to_bytes());
-            }
-            cert_bytes
-        }
-    }
-
-    /// Receive secret share after summing the secret input from each [`Contributor`] with
-    /// [`collect_secret_inputs`] and getting the `AggKeygenInput` from the coordinator.
-    ///
-    /// This also validates `agg_input`.
-    pub fn receive_secret_share<H, NG>(
-        schnorr: &Schnorr<H, NG>,
-        agg_input: &AggKeygenInput,
-        secret_share: SecretShare,
-    ) -> Result<PairedSecretShare<Normal, Zero>, ReceiveShareError>
-    where
-        H: Hash32,
-    {
-        for (key_contrib, pop) in &agg_input.key_contrib {
-            let (first_coeff_even_y, _) = key_contrib.into_point_with_even_y();
-            if !schnorr.verify(&first_coeff_even_y, Message::empty(), pop) {
-                return Err(ReceiveShareError::InvalidPop);
-            }
-        }
-
-        let shared_key = agg_input.shared_key();
-
-        let paired_secret_share = shared_key
-            .pair_secret_share(secret_share)
-            .ok_or(ReceiveShareError::InvalidSecretShare)?;
-
-        Ok(paired_secret_share)
-    }
-
-    /// Collect the secret inputs from each [`Contributor`] destined for a particular party at `ShareIndex`.
-    pub fn collect_secret_inputs(
-        my_index: ShareIndex,
-        secret_share_inputs: impl IntoIterator<Item = Scalar<Secret, Zero>>,
-    ) -> SecretShare {
-        let mut sum = s!(0);
-        for share in secret_share_inputs {
-            sum += share;
-        }
-
-        SecretShare {
-            index: my_index,
-            share: sum,
-        }
-    }
-
-    /// Simulate running a key generation with `simplepedpop`.
-    ///
-    /// This calls all the other functions defined in this module to get the whole job done on a
-    /// single computer by simulating all the other parties.
-    pub fn simulate_keygen<H, NG>(
-        schnorr: &Schnorr<H, NG>,
-        threshold: u32,
-        n_receivers: u32,
-        n_generators: u32,
-        rng: &mut impl rand_core::RngCore,
-    ) -> (SharedKey<Normal>, Vec<PairedSecretShare<Normal>>)
-    where
-        H: Hash32,
-        NG: NonceGen,
-    {
-        let share_receivers = (1..=n_receivers)
-            .map(|i| ShareIndex::from(NonZeroU32::new(i).unwrap()))
-            .collect::<BTreeSet<_>>();
-
-        let mut aggregator = Coordinator::new(threshold, n_generators);
-        let mut contributors = vec![];
-        let mut secret_inputs = BTreeMap::<ShareIndex, Vec<Scalar<Secret, Zero>>>::default();
-
-        for i in 0..n_generators {
-            let (contributor, to_coordinator, shares) =
-                Contributor::gen_keygen_input(schnorr, threshold, &share_receivers, i, rng);
-
-            contributors.push(contributor);
-            aggregator.add_input(schnorr, i, to_coordinator).unwrap();
-
-            for (receiver_index, share) in shares {
-                secret_inputs.entry(receiver_index).or_default().push(share);
-            }
-        }
-
-        let agg_input = aggregator.finish().unwrap();
-
-        for contributor in contributors {
-            contributor.verify_agg_input(&agg_input).unwrap();
-        }
-
-        let mut paired_shares = vec![];
-
-        for receiver in share_receivers {
-            let secret_share =
-                collect_secret_inputs(receiver, secret_inputs.remove(&receiver).unwrap());
-            let paired_share = receive_secret_share(schnorr, &agg_input, secret_share).unwrap();
-            paired_shares.push(paired_share.non_zero().unwrap());
-        }
-
-        (agg_input.shared_key().non_zero().unwrap(), paired_shares)
-    }
-
-    /// The input the contributor provided has been manipulated
-    #[derive(Clone, Copy, Debug, PartialEq)]
-    pub struct ContributionDidntMatch;
-
-    impl core::fmt::Display for ContributionDidntMatch {
-        fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-            write!(
-                f,
-                "the contribution assigned to us was not what we contributed"
-            )
-        }
-    }
-
-    #[cfg(feature = "std")]
-    impl std::error::Error for ContributionDidntMatch {}
-
-    /// The [`AggKeygenInput`] was invalid so a valid secret share couldn't be extracted.
-    #[derive(Clone, Copy, Debug, PartialEq)]
-    pub enum ReceiveShareError {
-        /// Invalid POP for one of the contributions
-        InvalidPop,
-        /// The secret share we got was invalid
-        InvalidSecretShare,
-    }
-
-    impl core::fmt::Display for ReceiveShareError {
-        fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-            write!(
-                f,
-                "{}",
-                match self {
-                    ReceiveShareError::InvalidPop => "Invalid POP for one of the contributions",
-                    ReceiveShareError::InvalidSecretShare =>
-                        "The share extracted from the key generation was invalid",
-                }
-            )
-        }
-    }
-}
-
-/// `encpedpop` is built on top of [`simplepedpop`] to add share encryption.
-///
-/// Each per recipient secret key is explicitly encrypted to each recipient and sent through the
-/// coordinator. This simplifies things a bit since all messages are to or from the coordinator. The
-/// coordinator also aggregates the ciphertexts so communication is reduced to linear in the number
-/// of participants.
-///
-/// The application still must figure out when all parties agree on the [`AggKeygenInput`] before
-/// using it.
-///
-/// [`AggKeygenInput`]: encpedpop::AggKeygenInput
-pub mod encpedpop {
-    use super::{simplepedpop, *};
-    use crate::frost::{Fingerprint, PairedSecretShare, SecretShare, ShareIndex, SharedKey};
-
-    /// A party that generates secret input to the key generation. You need at least one of these
-    /// and if at least one of these parties is honest then the final secret key will not be known by an
-    /// attacker (unless they obtain `t` shares!).
-    #[derive(Clone, Debug, PartialEq)]
-    #[cfg_attr(feature = "bincode", derive(bincode::Encode, bincode::Decode))]
-    #[cfg_attr(
-        feature = "serde",
-        derive(crate::fun::serde::Deserialize, crate::fun::serde::Serialize),
-        serde(crate = "crate::fun::serde")
-    )]
-    pub struct Contributor {
-        inner: simplepedpop::Contributor,
-    }
-
-    impl Contributor {
-        /// Generates the keygen input for a party at `my_index`. Note that `my_index`
-        /// has nothing to do with the "receiver" index (the `ShareIndex` of share receivers). If
-        /// there are `n` `KeyGenInputParty`s then each party must be assigned an index from `0` to `n-1`.
-        ///
-        /// This method return `Self` to retain the state of the protocol which is needded to verify
-        /// the aggregated input later on.
-        pub fn gen_keygen_input<H, NG>(
-            schnorr: &Schnorr<H, NG>,
-            threshold: u32,
-            receiver_encryption_keys: &BTreeMap<ShareIndex, Point>,
-            my_index: u32,
-            rng: &mut impl rand_core::RngCore,
-        ) -> (Self, KeygenInput)
-        where
-            H: Hash32,
-            NG: NonceGen,
-        {
-            let multi_nonce_keypair = KeyPair::<Normal>::new(Scalar::random(rng));
-
-            let share_receivers = receiver_encryption_keys.keys().cloned().collect();
-            let (inner_state, inner_keygen_input, mut shares) =
-                simplepedpop::Contributor::gen_keygen_input(
-                    schnorr,
-                    threshold,
-                    &share_receivers,
-                    my_index,
-                    rng,
-                );
-            let encryption_jobs = receiver_encryption_keys
-                .iter()
-                .map(|(receiver, encryption_key)| {
-                    (
-                        *receiver,
-                        (*encryption_key, shares.remove(receiver).unwrap()),
-                    )
-                })
-                .collect();
-            assert!(shares.is_empty());
-            let encrypted_shares = encrypt::<H>(encryption_jobs, multi_nonce_keypair);
-            let keygen_input = KeygenInput {
-                inner: inner_keygen_input,
-                encrypted_shares,
-                encryption_nonce: multi_nonce_keypair.public_key(),
-            };
-
-            (Contributor { inner: inner_state }, keygen_input)
-        }
-
-        /// Verifies that the coordinator has honestly included this party's input into the
-        /// aggregated input.
-        ///
-        /// This passing by itself doesn't mean that the key generation was successful. All
-        /// `Contributor`s must agree on this fact and all parties must have received the same
-        /// `AggKeygenInput` and validated it.
-        pub fn verify_agg_input(
-            self,
-            agg_keygen_input: &AggKeygenInput,
-        ) -> Result<(), simplepedpop::ContributionDidntMatch> {
-            self.inner.verify_agg_input(&agg_keygen_input.inner)?;
-            Ok(())
-        }
-    }
-
-    /// Key generation inputs after being aggregated by the coordinator
-    #[derive(Clone, Debug, PartialEq)]
-    #[cfg_attr(feature = "bincode", derive(bincode::Encode, bincode::Decode))]
-    #[cfg_attr(
-        feature = "serde",
-        derive(crate::fun::serde::Deserialize, crate::fun::serde::Serialize),
-        serde(crate = "crate::fun::serde")
-    )]
-    pub struct AggKeygenInput {
-        inner: simplepedpop::AggKeygenInput,
-        encrypted_shares: BTreeMap<ShareIndex, (Point, Scalar<Public, Zero>)>,
-        encryption_nonces: Vec<Point>,
-    }
-
-    impl AggKeygenInput {
-        /// Gets the `SharedKey` that this aggregated input produces.
-        ///
-        /// ## Security
-        ///
-        /// ⚠ Just because you can call this doesn't mean you can use the `SharedKey` securely yet!
-        ///
-        /// You have to have checked that all parties (contributors and receivers) think it's valid
-        /// *and* have the same copy first.
-        pub fn shared_key(&self) -> SharedKey<Normal, Zero> {
-            self.inner.shared_key()
-        }
-
-        /// The *certification* bytes. Checking all parties have the same output of this function is
-        /// enough to check they have the same `AggKeygenInput`.
-        pub fn cert_bytes(&self) -> Vec<u8> {
-            let mut cert_bytes = self.inner.cert_bytes();
-            cert_bytes.extend((self.encryption_nonces.len() as u32).to_be_bytes());
-            cert_bytes.extend(
-                self.encryption_nonces
-                    .iter()
-                    .flat_map(|nonce| nonce.to_bytes()),
-            );
-            cert_bytes.extend((self.encrypted_shares.len() as u32).to_be_bytes());
-            for (party_index, (encryption_key, encrypted_share)) in &self.encrypted_shares {
-                cert_bytes.extend(party_index.to_bytes());
-                cert_bytes.extend(encryption_key.to_bytes());
-                cert_bytes.extend(encrypted_share.to_bytes());
-            }
-            cert_bytes
-        }
-
-        /// Get the encryption key for every party
-        pub fn encryption_keys(&self) -> impl Iterator<Item = (ShareIndex, Point)> + '_ {
-            self.encrypted_shares
-                .iter()
-                .map(|(party_index, (ek, _))| (*party_index, *ek))
-        }
-
-        /// Certify the `AggKeygenInput`. If all parties certify this then the keygen was
-        /// successful.
-        pub fn certify<H, NG>(
-            &self,
-            schnorr: &Schnorr<H, NG>,
-            keypair: &KeyPair<EvenY>,
-        ) -> Signature
-        where
-            H: Hash32,
-            NG: NonceGen,
-        {
-            schnorr.sign(
-                keypair,
-                Message::new("BIP DKG/cert", self.cert_bytes().as_ref()),
-            )
-        }
-
-        /// Verify that another party has certified the keygen. If you collect certifications from
-        /// all parties then the keygen was successful
-        pub fn verify_cert<H: Hash32, NG>(
-            &self,
-            schnorr: &Schnorr<H, NG>,
-            cert_key: Point<EvenY>,
-            signature: Signature,
-        ) -> bool {
-            schnorr.verify(
-                &cert_key,
-                Message::new("BIP DKG/cert", self.cert_bytes().as_ref()),
-                &signature,
-            )
-        }
-
-        /// Recover a share with the decryption key from the `AggKeygenInput`.
-        pub fn recover_share<H: Hash32>(
-            &self,
-            share_index: ShareIndex,
-            keypair: &KeyPair,
-        ) -> Result<PairedSecretShare, &'static str> {
-            let (expected_public_key, agg_ciphertext) = self
-                .encrypted_shares
-                .get(&share_index)
-                .ok_or("No party at party_index existed")?;
-
-            if *expected_public_key != keypair.public_key() {
-                return Err("this isn't the right encryption keypair for this share");
-            }
-            let secret_share = decrypt::<H>(
-                share_index,
-                keypair,
-                &self.encryption_nonces,
-                *agg_ciphertext,
-            );
-
-            let paired_secret_share = self
-                .shared_key()
-                .pair_secret_share(SecretShare {
-                    index: share_index,
-                    share: secret_share,
-                })
-                .ok_or("the secret share recovered didn't match what was expected")?;
-
-            paired_secret_share
-                .non_zero()
-                .ok_or("the shared secret was zero")
-        }
-
-        /// Grinds all polynomial coefficients to achieve a fingerprint by rejection sampling.
-        /// This is meant to be run by the coordinator.
-        ///
-        /// This method modifies each non-constant coefficient of the polynomial through group
-        /// addition until the running hash (computed by sequentially incorporating coefficients)
-        /// has the required number of leading zero bits at each step.
-        ///
-        /// The fingerprint is tied to the specific public key by including it in the hash.
-        ///
-        /// ## Parameters
-        /// - `fingerprint`: The fingerprint configuration specifying difficulty and tag
-        ///
-        /// ## Security
-        /// This modification does not affect the security of the DKG as it only modifies
-        /// non-constant polynomial coefficients which are malleable by the coordinator.
-        pub fn grind_fingerprint<H: Hash32>(&mut self, fingerprint: Fingerprint) {
-            if self.inner.agg_poly.is_empty() {
-                return;
-            }
-
-            // Get the public key (first coefficient) by aggregating key contributions
-            let public_key = self
-                .inner
-                .key_contrib
-                .iter()
-                .fold(Point::zero(), |agg, (point, _)| g!(agg + point))
-                .normalize();
-
-            // Start with hash including the tag with length prefix
-            let mut hash_state = H::default();
-            if !fingerprint.tag.is_empty() {
-                hash_state = hash_state.add([fingerprint.tag.len() as u8]);
-                hash_state = hash_state.add(fingerprint.tag.as_bytes());
-            }
-            hash_state = hash_state.add(public_key);
-
-            let mut tweaks = Vec::with_capacity(self.inner.agg_poly.len());
-
-            // Grind each coefficient in sequence, note that agg_poly only
-            // contains the non-constant coefficients.
-            for coeff_index in 0..self.inner.agg_poly.len() {
-                let mut total_tweak = Scalar::<Public, Zero>::zero();
-                let original_coeff = self.inner.agg_poly[coeff_index];
-                let mut current_coeff = original_coeff;
-
-                loop {
-                    // Clone the hash state from previous coefficients and add current one
-                    let hash = hash_state.clone().add(current_coeff);
-                    // Check if hash has required number of leading zero bits
-                    let hash_bytes: [u8; 32] = hash.clone().finalize_fixed().into();
-                    if Fingerprint::leading_zero_bits(&hash_bytes[..])
-                        >= fingerprint.bit_length as usize
-                    {
-                        // Update hash_state for next coefficient because the next coefficient hash
-                        // includes the current one.
-                        hash_state = hash;
-                        break;
-                    }
-
-                    // Add one more G to the coefficient
-                    total_tweak += s!(1);
-                    current_coeff = g!(current_coeff + G).normalize();
-                }
-
-                // Apply the final tweak to the polynomial coefficient
-                self.inner.agg_poly[coeff_index] = current_coeff;
-                tweaks.push(total_tweak);
-            }
-
-            // Apply all tweaks to encrypted shares at once
-            for (share_index, (_enc_key, encrypted_share)) in &mut self.encrypted_shares {
-                let mut power = *share_index;
-                for tweak in &tweaks {
-                    *encrypted_share += s!(tweak * power).public();
-                    power *= share_index;
-                }
-            }
-        }
-    }
-
-    /// Produced by [`Contributor::gen_keygen_input`]. This is sent from the each
-    /// `Contributor` to the *coordinator*.
-    #[cfg_attr(feature = "bincode", derive(bincode::Encode, bincode::Decode))]
-    #[cfg_attr(
-        feature = "serde",
-        derive(crate::fun::serde::Deserialize, crate::fun::serde::Serialize),
-        serde(crate = "crate::fun::serde")
-    )]
-    #[derive(Clone, Debug, PartialEq)]
-    pub struct KeygenInput {
-        /// The input from the inner protocol
-        pub inner: simplepedpop::KeygenInput,
-        /// The shares encrypted for each receiving party
-        pub encrypted_shares: BTreeMap<ShareIndex, Scalar<Public, Zero>>,
-        /// The multi-encryption nonce for the encryptions in `encrypted_shares`
-        pub encryption_nonce: Point,
-    }
-
-    /// Stores the state of the coordinator as it aggregates inputs from [`Contributor`]s.
-    #[derive(Clone, Debug, PartialEq)]
-    pub struct Coordinator {
-        inner: simplepedpop::Coordinator,
-        agg_encrypted_shares: BTreeMap<ShareIndex, (Point, Scalar<Public, Zero>)>,
-        encryption_nonces: Vec<Point>,
-    }
-
-    impl Coordinator {
-        /// Creates a new coordinator with:
-        ///
-        /// - `threshold`: of key we're trying to generate
-        /// - `n_contributors`: The number of [`Contributor`]s
-        /// - `receiver_encryption_keys`: The encryption keys of each of the share receivers.
-        pub fn new(
-            threshold: u32,
-            n_contribtors: u32,
-            receiver_encryption_keys: &BTreeMap<ShareIndex, Point>,
-        ) -> Self {
-            let agg_encrypted_shares = receiver_encryption_keys
-                .iter()
-                .map(|(&receiver, encryption_key)| (receiver, (*encryption_key, Scalar::zero())))
-                .collect();
-            Self {
-                inner: simplepedpop::Coordinator::new(threshold, n_contribtors),
-                agg_encrypted_shares,
-                encryption_nonces: Default::default(),
-            }
-        }
-
-        /// Adds an `input` from a [`Contributor`].
-        ///
-        /// Note verifying this is the correct input from the correct party is up to your application!
-        pub fn add_input<H: Hash32, NG>(
-            &mut self,
-            schnorr: &Schnorr<H, NG>,
-            from: u32,
-            input: KeygenInput,
-        ) -> Result<(), &'static str> {
-            if self.inner.is_finished() {
-                return Err("all inputs have already been collected");
-            }
-            let mut check_missing = self.agg_encrypted_shares.keys().collect::<BTreeSet<_>>();
-
-            for dest in input.encrypted_shares.keys() {
-                if !self.agg_encrypted_shares.contains_key(dest) {
-                    return Err("included share for unknown party");
-                }
-                check_missing.remove(dest);
-            }
-
-            if !check_missing.is_empty() {
-                return Err("didn't have share for all parties");
-            }
-
-            // ⚠ only do mutations after we're sure everything is OK
-            self.inner.add_input(schnorr, from, input.inner)?;
-
-            for (dest, encrypted_share_contrib) in input.encrypted_shares {
-                let agg_encrypted_share = &mut self.agg_encrypted_shares.get_mut(&dest).unwrap().1;
-                *agg_encrypted_share += encrypted_share_contrib;
-            }
-
-            self.encryption_nonces.push(input.encryption_nonce);
-
-            Ok(())
-        }
-
-        /// Which [`Contributor`]s are we missing input from.
-        pub fn missing_from(&self) -> BTreeSet<u32> {
-            self.inner.missing_from()
-        }
-
-        /// Has the coordinator received input from each [`Contributor`].
-        pub fn is_finished(&self) -> bool {
-            self.inner.is_finished()
-        }
-
-        /// Try and finish input aggregation step.
-        ///
-        /// Returns `None` if [`is_finished`] returns `false`.
-        ///
-        /// [`is_finished`]: Self::is_finished
-        pub fn finish(self) -> Option<AggKeygenInput> {
-            let inner = self.inner.finish()?;
-            Some(AggKeygenInput {
-                inner,
-                encrypted_shares: self.agg_encrypted_shares,
-                encryption_nonces: self.encryption_nonces,
-            })
-        }
-    }
-
-    /// Extract our secret share from the `AggKeygenInput`.
-    ///
-    /// This also validates `agg_input`.
-    pub fn receive_secret_share<H, NG>(
-        schnorr: &Schnorr<H, NG>,
-        my_index: ShareIndex,
-        keypair: &KeyPair,
-        agg_input: &AggKeygenInput,
-    ) -> Result<PairedSecretShare<Normal, Zero>, simplepedpop::ReceiveShareError>
-    where
-        H: Hash32,
-    {
-        let encrypted_share = agg_input
-            .encrypted_shares
-            .get(&my_index)
-            .map(|(_pk, share)| *share)
-            .unwrap_or_default();
-        let share_scalar = decrypt::<H>(
-            my_index,
-            keypair,
-            &agg_input.encryption_nonces,
-            encrypted_share,
-        );
-        let secret_share = SecretShare {
-            index: my_index,
-            share: share_scalar,
-        };
-        let paired_secret_share =
-            simplepedpop::receive_secret_share(schnorr, &agg_input.inner, secret_share)?;
-
-        Ok(paired_secret_share)
-    }
-
-    fn encrypt<H: Hash32>(
-        encryption_jobs: BTreeMap<ShareIndex, (Point, Scalar<Secret, Zero>)>,
-        multi_nonce_keypair: KeyPair<Normal>,
-    ) -> BTreeMap<ShareIndex, Scalar<Public, Zero>> {
-        encryption_jobs
-            .iter()
-            .map(|(dest, (encryption_key, share))| {
-                let dh_key = g!(multi_nonce_keypair.secret_key() * encryption_key).normalize();
-                // SPEC DEVIATION: Hash inputs are as defined in "Multi-recipient Encryption, Revisited" by Pinto et al.
-                let pad = Scalar::from_hash(H::default().add(dh_key).add(encryption_key).add(dest));
-                let payload = s!(pad + share).public();
-                (*dest, payload)
-            })
-            .collect()
-    }
-
-    fn decrypt<H: Hash32>(
-        my_index: ShareIndex,
-        keypair: &KeyPair<Normal>,
-        multi_nocnes: &[Point],
-        mut agg_ciphertext: Scalar<Public, Zero>,
-    ) -> Scalar<Secret, Zero> {
-        for nonce in multi_nocnes {
-            let dh_key = g!(keypair.secret_key() * nonce).normalize();
-            let pad = Scalar::from_hash(
-                H::default()
-                    .add(dh_key)
-                    .add(keypair.public_key())
-                    .add(my_index),
-            );
-            agg_ciphertext -= pad;
-        }
-        agg_ciphertext.secret()
-    }
-
-    /// Simulate running a key generation with `encpedpop`.
-    ///
-    /// This calls all the other functions defined in this module to get the whole job done on a
-    /// single computer by simulating all the other parties.
-    ///
-    /// A fingerprint can be provided to grind into the polynomial coefficients.
-    pub fn simulate_keygen<H, NG>(
-        schnorr: &Schnorr<H, NG>,
-        threshold: u32,
-        n_receivers: u32,
-        n_generators: u32,
-        fingerprint: Fingerprint,
-        rng: &mut impl rand_core::RngCore,
-    ) -> (SharedKey<Normal>, Vec<PairedSecretShare<Normal>>)
-    where
-        H: Hash32,
-        NG: NonceGen,
-    {
-        let share_receivers = (1..=n_receivers)
-            .map(|i| Scalar::from(i).non_zero().unwrap())
-            .collect::<BTreeSet<_>>();
-
-        let receiver_enckeys = share_receivers
-            .iter()
-            .cloned()
-            .map(|party_index| (party_index, KeyPair::new(Scalar::random(rng))))
-            .collect::<BTreeMap<_, _>>();
-
-        let public_receiver_enckeys = receiver_enckeys
-            .iter()
-            .map(|(party_index, enckeypair)| (*party_index, enckeypair.public_key()))
-            .collect::<BTreeMap<ShareIndex, Point>>();
-
-        let (contributors, to_coordinator_messages): (Vec<Contributor>, Vec<KeygenInput>) = (0
-            ..n_generators)
-            .map(|i| {
-                Contributor::gen_keygen_input(schnorr, threshold, &public_receiver_enckeys, i, rng)
-            })
-            .unzip();
-
-        let mut aggregator = Coordinator::new(threshold, n_generators, &public_receiver_enckeys);
-
-        for (i, to_coordinator_message) in to_coordinator_messages.into_iter().enumerate() {
-            aggregator
-                .add_input(schnorr, i as u32, to_coordinator_message)
-                .unwrap();
-        }
-
-        let mut agg_input = aggregator.finish().unwrap();
-
-        // Apply fingerprint grinding
-        agg_input.grind_fingerprint::<H>(fingerprint);
-
-        for contributor in contributors {
-            contributor.verify_agg_input(&agg_input).unwrap();
-        }
-
-        let mut paired_secret_shares = vec![];
-        for (party_index, enckey) in receiver_enckeys {
-            let paired_secret_share =
-                receive_secret_share(schnorr, party_index, &enckey, &agg_input).unwrap();
-            paired_secret_shares.push(paired_secret_share.non_zero().unwrap());
-        }
-
-        let shared_key = agg_input.shared_key().non_zero().unwrap();
-        (shared_key, paired_secret_shares)
-    }
-}
-
-/// `certpedpop` is built on top of [`encpedpop`] to add certification of the outcome.
-///
-/// In [`encpedpop`] and [`simplepedpop`] it's left up to the application to figure out whether all
-/// the parties agree on the `AggKeygenInput`. In `certpedpop` the relevant methods return
-/// certification signatures on `AggKeygenInput` once they've been validated so they can be
-/// collected by the share receivers. Once the share receivers have got all the certificates they
-/// can finally output the key.
-///
-/// Certificates are collected from other share receivers as well as `Contributor`s.
-pub mod certpedpop {
-    use crate::frost::shared_key::Fingerprint;
-
-    use super::*;
-
-    /// A party that generates secret input to the key generation. You need at least one of these
-    /// and if at least one of these parties is honest then the final secret key will not be known by an
-    /// attacker (unless they obtain `t` shares!).
-    pub struct Contributor {
-        inner: encpedpop::Contributor,
-    }
-
-    /// Produced by [`Contributor::gen_keygen_input`]. This is sent from the each
-    /// `Contributor` to the *coordinator*.
-    pub type KeygenInput = encpedpop::KeygenInput;
-    /// Key generation inputs after being aggregated by the coordinator
-    pub type AggKeygenInput = encpedpop::AggKeygenInput;
-    /// A certificate containing signatures or proofs from certifying parties
-    #[derive(Clone, Debug)]
-    pub struct Certificate<Sig>(BTreeMap<Point, Sig>);
-
-    impl<Sig> Default for Certificate<Sig> {
-        fn default() -> Self {
-            Self(BTreeMap::new())
-        }
-    }
-
-    impl<Sig> Certificate<Sig> {
-        /// Create a new empty certificate
-        pub fn new() -> Self {
-            Self::default()
-        }
-
-        /// Insert a signature/proof for a public key
-        pub fn insert(&mut self, key: Point, sig: Sig) {
-            self.0.insert(key, sig);
-        }
-
-        /// Get the signature/proof for a public key
-        pub fn get(&self, key: &Point) -> Option<&Sig> {
-            self.0.get(key)
-        }
-
-        /// Iterate over all entries in the certificate
-        pub fn iter(&self) -> impl Iterator<Item = (&Point, &Sig)> {
-            self.0.iter()
-        }
-    }
-
-    impl Contributor {
-        /// Generates the keygen input for a party at `my_index`. Note that `my_index`
-        /// has nothing to do with the "receiver" index (the `ShareIndex` of share receivers). If
-        /// there are `n` `KeyGenInputParty`s then each party must be assigned an index from `0` to `n-1`.
-        ///
-        /// This method return `Self` to retain the state of the protocol which is needded to verify
-        /// the aggregated input later on.
-        pub fn gen_keygen_input<H: Hash32, NG: NonceGen>(
-            schnorr: &Schnorr<H, NG>,
-            threshold: u32,
-            receiver_encryption_keys: &BTreeMap<ShareIndex, Point>,
-            my_index: u32,
-            rng: &mut impl rand_core::RngCore,
-        ) -> (Self, KeygenInput) {
-            let (inner, message) = encpedpop::Contributor::gen_keygen_input(
-                schnorr,
-                threshold,
-                receiver_encryption_keys,
-                my_index,
-                rng,
-            );
-            (Self { inner }, message)
-        }
-
-        /// Verifies that the coordinator has honestly included this party's input into the
-        /// aggregated input and returns a certification signature to that effect.
-        ///
-        /// This passing by itself doesn't mean that the key generation was successful. You must
-        /// first collect the signatures from all the certifying parties (contributors and share
-        /// receivers).
-        pub fn verify_agg_input<S: CertificationScheme>(
-            self,
-            cert_scheme: &S,
-            agg_keygen_input: &AggKeygenInput,
-            cert_keypair: &KeyPair,
-        ) -> Result<S::Signature, simplepedpop::ContributionDidntMatch> {
-            self.inner.verify_agg_input(agg_keygen_input)?;
-            let sig = cert_scheme.certify(cert_keypair, agg_keygen_input);
-            Ok(sig)
-        }
-    }
-
-    /// A key generation session that has been certified by each certifying party (contributors and share receivers).
-    #[derive(Clone, Debug)]
-    pub struct CertifiedKeygen<S: CertificationScheme> {
-        input: AggKeygenInput,
-        certificate: Certificate<S::Signature>,
-        /// The outputs from successful verification, indexed by certifying party's public key
-        outputs: BTreeMap<Point, S::Output>,
-    }
-
-    impl<S: CertificationScheme> CertifiedKeygen<S> {
-        /// Recover a share from a certified key generation with the decryption key.
-        ///
-        /// This checks that the `keypair` has signed the key generation first.
-        pub fn recover_share<H: Hash32>(
-            &self,
-            cert_scheme: &S,
-            share_index: ShareIndex,
-            keypair: KeyPair,
-        ) -> Result<PairedSecretShare, &'static str> {
-            let cert_key = keypair.public_key();
-            let my_cert = self
-                .certificate
-                .get(&cert_key)
-                .ok_or("I haven't certified this keygen")?;
-            if cert_scheme
-                .verify_cert(cert_key, &self.input, my_cert)
-                .is_none()
-            {
-                return Err("my certification was invalid");
-            }
-            self.input.recover_share::<H>(share_index, &keypair)
-        }
-
-        /// Gets the inner `encpedpop::AggKeygenInput`.
-        pub fn inner(&self) -> &AggKeygenInput {
-            &self.input
-        }
-
-        /// Gets the certificate.
-        pub fn certificate(&self) -> &Certificate<S::Signature> {
-            &self.certificate
-        }
-
-        /// Gets the verification outputs.
-        pub fn outputs(&self) -> &BTreeMap<Point, S::Output> {
-            &self.outputs
-        }
-    }
-
-    #[cfg(feature = "vrf_cert_keygen")]
-    impl CertifiedKeygen<vrf_cert::VrfCertifier> {
-        /// Compute a randomness beacon from the VRF outputs
-        ///
-        /// This function hashes all the VRF gamma points together to produce
-        /// unpredictable randomness that no single party could have controlled
-        /// (as long as at least one party is honest).
-        pub fn compute_randomness_beacon(&self) -> [u8; 32] {
-            use sha2::{Digest, Sha256};
-
-            let mut hasher = Sha256::new();
-
-            // BTreeMap already maintains sorted order by key
-            for output in self.outputs.values() {
-                hasher.update(output.gamma.to_bytes());
-            }
-
-            hasher.finalize().into()
-        }
-    }
-
-    pub use encpedpop::Coordinator;
-
-    /// Stores the state of share recipient who first receives their share and then waits to get
-    /// signatures from all the certifying parties on the keygeneration before accepting it.
-    pub struct SecretShareReceiver {
-        paired_secret_share: PairedSecretShare<Normal, Zero>,
-        agg_input: AggKeygenInput,
-    }
-
-    impl SecretShareReceiver {
-        /// Extract your `keypair` and certify the key generation. Before you actually
-        /// can use the share you must call [`finalize`] with a completed certificate.
-        ///
-        /// [`finalize`]: Self::finalize
-        pub fn receive_secret_share<H, NG, S>(
-            schnorr: &Schnorr<H, NG>,
-            cert_scheme: &S,
-            my_index: ShareIndex,
-            keypair: &KeyPair,
-            agg_input: &AggKeygenInput,
-        ) -> Result<(Self, S::Signature), simplepedpop::ReceiveShareError>
-        where
-            H: Hash32,
-            NG: NonceGen,
-            S: CertificationScheme,
-        {
-            let paired_secret_share =
-                encpedpop::receive_secret_share(schnorr, my_index, keypair, agg_input)?;
-            let sig = cert_scheme.certify(keypair, agg_input);
-            let self_ = Self {
-                paired_secret_share,
-                agg_input: agg_input.clone(),
-            };
-            Ok((self_, sig))
-        }
-
-        /// Check the certificate contains a signature from each certifying party.
-        ///
-        /// By default every share receiver is a certifying party but you must also get
-        /// certifications from the [`Contributor`]s for security. Their keys are passed in as
-        /// `contributor_keys`.
-        pub fn finalize<S: CertificationScheme>(
-            self,
-            cert_scheme: &S,
-            certificate: Certificate<S::Signature>,
-            contributor_keys: &[Point],
-        ) -> Result<(CertifiedKeygen<S>, PairedSecretShare<Normal, Zero>), &'static str> {
-            let mut outputs = BTreeMap::new();
-            let cert_keys = self
-                .agg_input
-                .encryption_keys()
-                .map(|(_, encryption_key)| encryption_key)
-                .chain(contributor_keys.iter().cloned());
-            for cert_key in cert_keys {
-                match certificate.get(&cert_key) {
-                    Some(sig) => match cert_scheme.verify_cert(cert_key, &self.agg_input, sig) {
-                        Some(output) => {
-                            outputs.insert(cert_key, output);
-                        }
-                        None => return Err("certification signature was invalid"),
-                    },
-                    None => return Err("missing certification signature"),
-                }
-            }
-
-            let certified_keygen = CertifiedKeygen {
-                input: self.agg_input,
-                certificate,
-                outputs,
-            };
-
-            Ok((certified_keygen, self.paired_secret_share))
-        }
-    }
-
-    /// Simulate running a key generation with `certpedpop`.
-    ///
-    /// This calls all the other functions defined in this module to get the whole job done on a
-    /// single computer by simulating all the other parties.
-    ///
-    /// A fingerprint can be provided to grind into the polynomial coefficients.
-    pub fn simulate_keygen<H: Hash32, NG: NonceGen, S: CertificationScheme>(
-        schnorr: &Schnorr<H, NG>,
-        cert_scheme: &S,
-        threshold: u32,
-        n_receivers: u32,
-        n_generators: u32,
-        fingerprint: Fingerprint,
-        rng: &mut impl rand_core::RngCore,
-    ) -> (
-        CertifiedKeygen<S>,
-        Vec<(PairedSecretShare<Normal>, KeyPair)>,
-    ) {
-        let share_receivers = (1..=n_receivers)
-            .map(|i| Scalar::from(i).non_zero().unwrap())
-            .collect::<BTreeSet<_>>();
-
-        let mut receiver_enckeys = share_receivers
-            .iter()
-            .cloned()
-            .map(|party_index| (party_index, KeyPair::new(Scalar::random(rng))))
-            .collect::<BTreeMap<_, _>>();
-
-        let public_receiver_enckeys = receiver_enckeys
-            .iter()
-            .map(|(party_index, enckeypair)| (*party_index, enckeypair.public_key()))
-            .collect::<BTreeMap<ShareIndex, Point>>();
-
-        let (contributors, to_coordinator_messages): (Vec<Contributor>, Vec<KeygenInput>) = (0
-            ..n_generators)
-            .map(|i| {
-                Contributor::gen_keygen_input(schnorr, threshold, &public_receiver_enckeys, i, rng)
-            })
-            .unzip();
-
-        let contributor_keys = (0..n_generators)
-            .map(|_| KeyPair::new(Scalar::random(rng)))
-            .collect::<Vec<_>>();
-        let contributor_public_keys = contributor_keys
-            .iter()
-            .map(|kp| kp.public_key())
-            .collect::<Vec<_>>();
-
-        let mut aggregator = Coordinator::new(threshold, n_generators, &public_receiver_enckeys);
-
-        for (i, to_coordinator_message) in to_coordinator_messages.into_iter().enumerate() {
-            aggregator
-                .add_input(schnorr, i as u32, to_coordinator_message)
-                .unwrap();
-        }
-
-        let mut agg_input = aggregator.finish().unwrap();
-
-        // Apply fingerprint grinding
-        agg_input.grind_fingerprint::<H>(fingerprint);
-        let mut certificate = Certificate::new();
-
-        for (contributor, keypair) in contributors.into_iter().zip(contributor_keys.iter()) {
-            let sig = contributor
-                .verify_agg_input(cert_scheme, &agg_input, keypair)
-                .unwrap();
-            certificate.insert(keypair.public_key(), sig);
-        }
-
-        let mut paired_secret_shares = vec![];
-        let mut share_receivers = vec![];
-        for (party_index, enckey) in &receiver_enckeys {
-            let (share_receiver, cert) = SecretShareReceiver::receive_secret_share(
-                schnorr,
-                cert_scheme,
-                *party_index,
-                enckey,
-                &agg_input,
-            )
-            .unwrap();
-            certificate.insert(enckey.public_key(), cert);
-            share_receivers.push(share_receiver);
-        }
-
-        // Collect outputs by verifying all certificates
-        let mut outputs = BTreeMap::new();
-        for (key, sig) in certificate.iter() {
-            if let Some(output) = cert_scheme.verify_cert(*key, &agg_input, sig) {
-                outputs.insert(*key, output);
-            }
-        }
-
-        let certified_keygen = CertifiedKeygen {
-            input: agg_input.clone(),
-            certificate: certificate.clone(),
-            outputs,
-        };
-
-        for share_receiver in share_receivers {
-            let (_certified, paired_secret_share) = share_receiver
-                .finalize(cert_scheme, certificate.clone(), &contributor_public_keys)
-                .unwrap();
-            paired_secret_shares.push((
-                paired_secret_share.non_zero().unwrap(),
-                receiver_enckeys
-                    .remove(&paired_secret_share.index())
-                    .unwrap(),
-            ));
-        }
-
-        (certified_keygen, paired_secret_shares)
-    }
-
-    /// There was a problem with the keygen certificate so the key generation can't be trusted.
-    #[derive(Clone, Debug, Copy, PartialEq)]
-    pub enum CertificateError {
-        /// A certificate was invalid
-        InvalidCert {
-            /// The key that had the invalid cert
-            key: Point,
-        },
-        /// A certificate was missing
-        Missing {
-            /// They key whose cert was missing
-            key: Point,
-        },
-    }
-
-    impl core::fmt::Display for CertificateError {
-        fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
-            match self {
-                CertificateError::InvalidCert { key } => {
-                    write!(f, "certificate for key {key} was invalid")
-                }
-                CertificateError::Missing { key } => {
-                    write!(f, "certificate for key {key} was missing")
-                }
-            }
-        }
-    }
-
-    #[cfg(feature = "std")]
-    impl std::error::Error for CertificateError {}
-}
-
 #[cfg(test)]
 mod test {
+    use crate::frost::Fingerprint;
+
     use super::*;
     use proptest::{
         prelude::*,
